refactor: wire the upstream proxy explicitly via NewUpstreamProxy
NewHandler silently fell back to a credential-injecting reverse proxy
when next was unset, so every embedder forwarded unmatched requests to
the upstream whether or not that was wanted, and the proxy could not
be built or configured on its own. The proxy now lives in proxy.go as
NewUpstreamProxy(upstream, token) and callers opt in through WithNext;
cmd/xetd and the test fixtures wire it explicitly, and an unset next
answers 404 instead.
